+++ src/LaravelVandarServiceProvider.php
@@ -11,8 +11,10 @@ use Zarbinco\LaravelVandar\Contracts\TokenStore;
 use Zarbinco\LaravelVandar\Http\PendingVandarRequest;
 use Zarbinco\LaravelVandar\Http\VandarClient;
 use Zarbinco\LaravelVandar\Resources\BusinessResource;
+use Zarbinco\LaravelVandar\Resources\CardResource;
 use Zarbinco\LaravelVandar\Resources\CustomerFieldResource;
 use Zarbinco\LaravelVandar\Resources\CustomerResource;
+use Zarbinco\LaravelVandar\Resources\IbanResource;
 use Zarbinco\LaravelVandar\Resources\RawResource;
 use Zarbinco\LaravelVandar\Support\BusinessResolver;
 use Zarbinco\LaravelVandar\Token\CacheTokenStore;
@@ -37,6 +39,8 @@ final class LaravelVandarServiceProvider extends ServiceProvider
         $this->app->singleton(BusinessResolver::class);
         $this->app->singleton(BusinessResource::class);
         $this->app->singleton(CustomerFieldResource::class);
+        $this->app->singleton(CardResource::class);
+        $this->app->singleton(IbanResource::class);
         $this->app->singleton(CustomerResource::class);
 
         $this->app->singleton('vandar', fn ($app): LaravelVandar => new LaravelVandar(
@@ -46,6 +50,8 @@ final class LaravelVandarServiceProvider extends ServiceProvider
             $app->make(RawResource::class),
             $app->make(BusinessResource::class),
             $app->make(CustomerResource::class),
+            $app->make(CardResource::class),
+            $app->make(IbanResource::class),
         ));
         $this->app->alias('vandar', LaravelVandar::class);
     }